Check both positive and negative errors, and report consistently.
git-svn-id: https://icl.cs.utk.edu/svn/magma/trunk@2663 02bb5f2a-90ca-11de-bb24-8aedcc835bc8

+++ testing/testing_zgeqp3.cpp
@@ -91,7 +91,7 @@ int main( int argc, char** argv)
                 cpu_time = magma_wtime() - cpu_time;
                 cpu_perf = gflops / cpu_time;
                 if (info != 0)
-                    printf("lapack_zgeqp3 returned error %d.\n", info);
+                    printf("lapack_zgeqp3 returned error %d.\n", (int) info);
             }
             
             /* ====================================================================
@@ -110,7 +110,7 @@ int main( int argc, char** argv)
             gpu_time = magma_wtime() - gpu_time;
             gpu_perf = gflops / gpu_time;
             if (info != 0)
-                printf("magma_zgeqp3 returned error %d.\n", info);
+                printf("magma_zgeqp3 returned error %d.\n", (int) info);
             
             /* =====================================================================
                Check the result
